05/26

+++ src/com/javaex/ex03/Ex06.java
@@ -7,7 +7,7 @@ public class Ex06 {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
 		System.out.println("점수를 입력하세요.");
-		int score = sc.nextInt();
+		long score = sc.nextLong();
 
 		if (score % 3 == 0) {
 			System.out.println(score + "은(는) 3의 배수 입니다.");
